+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/Run.java
@@ -47,6 +47,7 @@ import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
+import javax.persistence.OneToOne;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
@@ -61,6 +62,7 @@ import uk.ac.bbsrc.tgac.miso.core.data.impl.RunPosition;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.UserImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.changelog.RunChangeLog;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.kit.KitDescriptor;
+import uk.ac.bbsrc.tgac.miso.core.data.impl.view.RunProjectView;
 import uk.ac.bbsrc.tgac.miso.core.data.type.HealthType;
 import uk.ac.bbsrc.tgac.miso.core.data.type.PlatformType;
 
@@ -164,6 +166,10 @@ public abstract class Run
   @ManyToOne(targetEntity = UserImpl.class)
   @JoinColumn(name = "dataApproverId")
   private User dataApprover;
+
+  @OneToOne
+  @JoinColumn(name = "runId", updatable = false, insertable = false)
+  private RunProjectView runProjectView;
 
   /**
    * Construct a new Run with a default empty SecurityProfile
@@ -470,6 +476,18 @@ public abstract class Run
 
   public void setDataApprover(User dataApprover) {
     this.dataApprover = dataApprover;
+  }
+
+  public RunProjectView getRunProjectView() {
+    return runProjectView;
+  }
+
+  public void setRunProjectView(RunProjectView runProjectView) {
+    this.runProjectView = runProjectView;
+  }
+
+  public String getProjectsLabel() {
+    return runProjectView == null ? null : runProjectView.getProjects();
   }
 
 }